Brand new codec.
